Add item count for grids (#595)
* work in progress

* more wip

* wip

* fix selection

* wip

* add count to rawdata tab

* remove unused import

* Fix measurement count within tab

* update count in sample view, remove experiment name

* address review comments

* fix filter/select issue, add all-select to raw data tab

* kill vite

* add comment

* remove remains of tabsheet

* review changes

* add equals and hashcode

* remove equals and hashcode

* clean up code and add comments

* simplify

---------

+++ project-management/src/main/java/life/qbic/projectmanagement/application/dataset/RawDataLookupService.java
@@ -47,7 +47,6 @@ public class RawDataLookupService {
     return new ArrayList<>(dataList);
   }
 
-
   /**
    * Provides the count of the registered measurements for the provided measurementIds
    *
@@ -55,7 +54,7 @@ public class RawDataLookupService {
    *                  determined
    * @return number of raw data for all domains associated with the provided measurementIds
    */
-  public long countRawDataByMeasurementCodes(Collection<MeasurementCode> measurementCodes) {
+  public int countRawDataByMeasurementCodes(Collection<MeasurementCode> measurementCodes) {
     return rawDataLookup.countRawDataByMeasurementIds(measurementCodes);
   }
